+++ src/routes/StudentRoutes.tsx
@@ -1,4 +1,4 @@
-import {Route, Routes} from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import Application from "@/pages/Student/Internship/Application";
 import EnterResult from "@/pages/Student/Internship/EnterResult";
 import SwTestEnterResult from "@/pages/Student/SwTest/SwTestEnterResult";
@@ -9,6 +9,7 @@ import CompetitionContest from "@pages/Student/SwTalent/ CompetitionContest";
 import Thesis from "@pages/Student/SwTalent/ Thesis";
 import DKUStarTrack from "@/pages/Student/scholarship/DKUStarTrack";
 import SwConvergence from "@/pages/Student/scholarship/sw-convergence/Index";
+import SwRisingStar from "@pages/Student/scholarship/SwRisingStar";
 
 const StudentRoutes = () => {
     return (
@@ -23,8 +24,12 @@ const StudentRoutes = () => {
             <Route path="sw-test/enter-result" element={<SwTestEnterResult/>}/>
             <Route path="scholarship/dku-startrack" element={<DKUStarTrack/>}/>
             <Route path="scholarship/sw-convergence" element={<SwConvergence />}/>
+            <Route path="scholarship/sw-rising-star" element={<SwRisingStar />} />
         </Routes>
     );
 };
 
 export default StudentRoutes;
+
+
+
